+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -136,6 +136,10 @@ New checks
   Future libc++ will remove the extension (`D120996
   <https://reviews.llvm.org/D120996>`).
 
+- New :doc:`misc-homoglyph <clang-tidy/checks/misc-homoglyph>` check.
+
+  Detects confusable unicode identifiers.
+
 New check aliases
 ^^^^^^^^^^^^^^^^^